Safari-611.1.10/LayoutTests/mathml/mfrac-nested.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This test passes if it does not time out.</p>
<math display="block">
<mrow>
<mfrac>
    <mi>a</mi>
    <mfrac>
        <mi>b</mi>
        <mfrac>
            <mi>c</mi>
            <mfrac>
                <mi>d</mi>
                <mfrac>
                    <mi>e</mi>
                    <mfrac>
                        <mi>f</mi>
                        <mfrac>
                            <mi>g</mi>
                            <mfrac>
                                <mi>h</mi>
                                <mfrac>
                                    <mi>i</mi>
                                    <mo>j</mo>
                                </mfrac>
                            </mfrac>
                        </mfrac>
                    </mfrac>
                </mfrac>
            </mfrac>
        </mfrac>
    </mfrac>
</mfrac>
</mrow>
</math>
</body>
</html>
